Add support for multi-threading in CDC Server.
The number of threads is controlled by the property "cdcsdk.threads".
Default is 1

The commit introduces a new engine - MTEngine which is an almost perfect
replica of EmbeddedEngine. The difference is that, MTEngine calls
`connector.taskConfigs` with the right parameter. Similarly it starts a
task with the right task id. EmbeddedEngine has hard coded these values
and is a `public final` class. So it can be sub-classed.

Every thread runs an MTEngine which consists of a source and a sink.
Each thread gets the list of tablets and sorts them. Based on the taskId
of each thread, a separate group of tablets is chosen.

Manually tested. Useful log lines:

    Number of threads: 4
    ...
    Every task will log the following:

    maxTasks = 4
    taskId = 1