Add S3 Sink to write inserts to a AWS S3 Location  (#1)
S3 Sink writes streams of INSERT CDC records to S3. The Sink accepts the following parameters:

baseDir: A Base Directory to write
pattern: A sub-directory and filename pattern. The pattern is used to create new files when rollover is triggered.
flushSizeMB: The size in bytes when rollover to a new file is triggered.
flushRecords: The number of records when rollover to a new file is triggered.
flushSecs: The number of seconds after which a rollover to new file is triggered.
A new file is created when any of the above conditions are met.

For more details check docs: https://docs.yugabyte.com/preview/explore/change-data-capture/cdcsdk-server/